# Faster flamegraph generation