--- a/lib/stackprof/report.rb
+++ b/lib/stackprof/report.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'pp'
 require 'digest/md5'