This PR makes flamegraph generation significantly faster. The main change is to create cursor objects pointing into the raw buffer instead of slicing arrays out of the raw buffer.

**Before**
```
jhawthorn@guardian:~/src/stackprof (master ) [ruby 2.8.0]
$ time ruby -Ilib bin/stackprof --flamegraph stackprof-test-boot.dump > /dev/null
31.52s user 0.24s system 99% cpu 31.891 total
```

**After**
```
jhawthorn@guardian:~/src/stackprof (faster_reporting ) [ruby 2.8.0]
$ time ruby -Ilib bin/stackprof --flamegraph stackprof-test-boot.dump > /dev/null
6.69s user 0.16s system 99% cpu 6.875 total
```


I found this using stackprof on its own reporting 😆 ♻️

``` ruby
# frozen_string_literal: true

require "stackprof"

report = StackProf::Report.new(Marshal.load(IO.binread(ARGV[0])))

output = File.open("/dev/null", "wb")
StackProf.run(mode: :wall, out: 'stackprof-stackprof.dump') do
  report.print_timeline_flamegraph(output)
end
```