@tmm1 2020-07-05
This is really clever. Great use of stackprof!